블록체인 라이브러리에서 안드로이드 의존성 제외함 - 코루틴은 어차피 안드로이드 아닌 것에서도 돌아가니 유지함 - 안드로이드가 아닌 다른 코틀린 환경에서 돌려봐야겠음
블록체인 난이도 구현 추가
- 난이도라는게 너무 많이 채굴하지 않도록 하는 것인데...
- 프라이빗 블록체인을 지향한다면 굳이 없어도 되는것
- 추후에는 난이도는 최소한만 하던지 선택을 하던지 지정을 할까 싶음
- 그래도 블록체인 구현을 배우니깐 난이도 조정은 이렇게 한다라고 구현하고 싶긴했음

dispatcher사용 안함
- 디스패쳐 사용을 하니, 메인 스레드를 잡아먹더라
- 비동기, 스레드, 코루틴 연구가 더 필요하다
- 결국엔 전용 스레드 만드는걸로 다시 복귀함

+++ BlockChain/src/main/java/cuisine/de/lapin/library/blockchain/utils/BlockChainUtils.kt
@@ -15,7 +15,7 @@ internal fun createBlock(
     previousHash: String,
     height: UInt,
     timeStamp: Long,
-    difficulty: Int
+    difficulty: UInt
 ): Block {
     var nonce: UInt = INIT_NONCE
     var payload = ""
@@ -42,8 +42,8 @@ internal fun createBlock(
     )
 }
 
-private fun getStartZeros(difficulty: Int): String {
-    return ZERO.repeat(difficulty)
+private fun getStartZeros(difficulty: UInt): String {
+    return ZERO.repeat(difficulty.toInt())
 }
 
 internal fun getPayload(block: Block): String {
@@ -55,7 +55,7 @@ internal fun getPayload(
     previousHash: String,
     height: UInt,
     timeStamp: Long,
-    difficulty: Int,
+    difficulty: UInt,
     nonce: UInt
 ): String {
     return "$content$previousHash$height$timeStamp$difficulty$nonce"
